Progress to setup titan configuration files for testing

+++ prepare_test.sh
@@ -1,7 +1,7 @@
 
 #1st param is the uImage to test
 uImagename=$1
-
+testsuitename=$2
 #Get the builds
 #TODO: use cURL to support multiple download methods
 wget --no-proxy -q  $TEST_BUILD_URL
@@ -22,9 +22,21 @@ fi
 
 wget --no-proxy -q $RESOURCES_URL"/Titan-software.tar.gz"
 tar -xvf Titan-software.tar.gz > /dev/null
+
+#TODO Implement a mechanism to download thirdparty
+#test descriptors and test lists which are not in kci-alpha
+if [ $testsuitename == "sanity_test" ]
+then
 mv Titan-software Titan
 rm -fr Titan/test_descriptors
 rm -f Titan/settings/titan-selectedtests.xml
 rm -f Titan/logs/*
-ls -al
 
+cp -r $KCI_ROOT/test_descriptors ./Titan/
+cp $KCI_ROOT/settings/user-settings.xml."$NODE_NAME"."$EXECUTOR_NUMBER" ./Titan/settings/user-settings.xml
+cp $KCI_ROOT/testsuite/$testsuitename Titan/settings/titan-selectedTests.xml
+
+else
+  echo "Not implemented "
+
+fi
